Removes output.go file
