+++ cloudfunctions/gce/createsnapshot/createsnapshot.go
@@ -21,7 +21,6 @@ import (
 	"strings"
 	"time"
 
-	"github.com/googlecloudplatform/security-response-automation/cloudfunctions/output"
 	"github.com/googlecloudplatform/security-response-automation/services"
 	"github.com/pkg/errors"
 	compute "google.golang.org/api/compute/v1"
@@ -60,6 +59,12 @@ type Services struct {
 	Resource *services.Resource
 }
 
+// Output contains the output of this function.
+type Output struct {
+	// DiskNames optionally contains the names of the disks copied to a target project.
+	DiskNames []string
+}
+
 // Execute creates a snapshot of an instance's disk.
 //
 // For a given supported finding pull each disk associated with the affected instance.
@@ -69,7 +74,7 @@ type Services struct {
 // In order for the snapshot to be create the service account must be granted the correct
 // role on the affected project. At this time this grant is defined per project but should
 // be changed to support folder and organization level grants.
-func Execute(ctx context.Context, values *Values, services *Services) (*output.Output, error) {
+func Execute(ctx context.Context, values *Values, services *Services) (*Output, error) {
 	log.Printf("listing disk names within instance %q, in zone %q and project %q", values.Instance, values.Zone, values.ProjectID)
 	disksCopied := []string{}
 	rule := strings.Replace(values.RuleName, "_", "-", -1)
@@ -130,7 +135,7 @@ func Execute(ctx context.Context, values *Values, services *Services) (*output.O
 		}
 	}
 	log.Printf("completed")
-	return &output.Output{DiskNames: disksCopied}, nil
+	return &Output{DiskNames: disksCopied}, nil
 }
 
 // canCreateSnapshot checks if we should create a snapshot along with a map of existing snapshots to be removed.
